Split lock parts into several files
