+++ rtl/src/main/scala/tensil/mutex/Comparable.scala
@@ -1,0 +1,8 @@
+package tensil.mutex
+
+import chisel3._
+
+trait Comparable[T <: Data] {
+  self: T =>
+  def ===(other: T): Bool
+}
