+++ rtl/src/main/scala/tensil/mutex/ConditionalRelease.scala
@@ -1,0 +1,8 @@
+package tensil.mutex
+
+import chisel3._
+
+trait ConditionalRelease[T <: Data] {
+  val delayRelease: UInt
+  val cond: T
+}
